+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - In-page anchor menus (for Navbar clarity)
 - Blog Section
 - Projects Section
+- My Other Channels Section
 ### Changed
 - Optimized some parts of JS Code
 - Converted Data into a static JSON file